fix: sanitize send_to_agent text + extract sanitizer to shared module
Macroscope HIGH: send_to_agent also sends text to terminals but
wasn't sanitized. Extracted sanitizeTerminalInput to src/sanitize.ts
(avoids circular dependency server<->agent-engine). Both send_input
and sendToAgent now sanitize.
